+++ scripts/gh-pages.js
@@ -6,7 +6,7 @@ ghpages.publish('gh-pages', function(err) {
   } else {
     console.log('*---------------------------------------------------------------------------*');
     console.log('|                                                                           |');
-    console.log('|     ✨⭐', '\x1b[32m', 'released at: ', '\x1b[4m\x1b[36m', 'https://litospayaso.github.io/open-cal', '\x1b[0m', '⭐✨         |');
+    console.log('|     ✨⭐', '\x1b[32m', 'released at: ', '\x1b[4m\x1b[36m', 'https://litospayaso.github.io/brote', '\x1b[0m', '⭐✨         |');
     console.log('|                                                                           |');
     console.log('*---------------------------------------------------------------------------*');
   }