+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # brote
 
+## 1.0.14
+
+### Patch Changes
+
+- Renaming project into brote and adding capacitor/status-bar to the apk build
+
 ## 1.0.13
 
 ### Patch Changes